+++ src/main/java/pageFactory/BackOfficeDashboardPage.java
@@ -1,4 +1,4 @@
-package pageObject;
+package pageFactory;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
