+++ src/main/java/pageFactory/BasePagePurchaseService.java
@@ -1,6 +1,5 @@
-package pageObject;
+package pageFactory;
 
-import org.openqa.selenium.JavascriptException;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
