+++ src/test/java/purchaseServiceTest/BaseTest.java
@@ -7,8 +7,8 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
-import pageObject.BackOfficeDashboardPage;
-import pageObject.LoginPage;
+import pageFactory.BackOfficeDashboardPage;
+import pageFactory.LoginPage;
 
 import static io.github.bonigarcia.wdm.WebDriverManager.chromedriver;
